third_party/WebKit/LayoutTests/http/tests/devtools/coverage/coverage-view-expected.txt

Tests the coverage list view after finishing recording in the Coverage view.

http://127.0.0.1:8000/devtools/coverage/resources/coverage.js JS (coarse) used: 407 unused: 161 total: 568
http://127.0.0.1:8000/devtools/coverage/resources/highlight-in-source.css CSS used: 67 unused: 142 total: 209
Reloading Page
Page reloaded.
http://127.0.0.1:8000/devtools/coverage/resources/coverage.js JS used: 354 unused: 214 total: 568
http://127.0.0.1:8000/devtools/coverage/resources/highlight-in-source.css CSS used: 67 unused: 142 total: 209
